feat: TableGroup 검증 로직 도메인으로 이관
- 도메인 테스트 추가
- 기존 service 테스트에서 검증 테스트 이관
