+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -10,7 +10,6 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import java.util.Arrays;
-import java.util.Collections;
 import kitchenpos.application.fixture.OrderTableFixtureFactory;
 import kitchenpos.application.fixture.TableGroupFixtureFactory;
 import kitchenpos.dao.OrderDao;
@@ -54,32 +53,6 @@ class TableGroupServiceTest {
     }
 
     @Test
-    @DisplayName("주문 테이블이 비어 있으면 테이블 그룹으로 지정 할 수 없다.")
-    void createFailWithEmptyTest() {
-
-        //given
-        TableGroup tableGroup = new TableGroup(1L, Collections.emptyList());
-
-        //when & then
-        assertThatThrownBy(
-                () -> tableGroupService.create(tableGroup)
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("주문 테이블이 2개보다 작으면 테이블 그룹으로 지정 할 수 없다.")
-    void createFailWithUnder2Test() {
-
-        //given
-        TableGroup tableGroup = new TableGroup(1L, Arrays.asList(주문_테이블1));
-
-        //when & then
-        assertThatThrownBy(
-                () -> tableGroupService.create(tableGroup)
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
     @DisplayName("주문 테이블들이 시스템에 등록 되어 있지 않으면 테이블 그룹은 지정 할 수 없다.")
     void createFailWithOrderTableNotExistTest() {
 
@@ -87,39 +60,6 @@ class TableGroupServiceTest {
         TableGroup tableGroup = new TableGroup(1L, Arrays.asList(주문_테이블1, 주문_테이블2));
         given(orderTableDao.findAllByIdIn(Arrays.asList(주문_테이블1.getId(), 주문_테이블2.getId()))).willReturn(
                 Arrays.asList(주문_테이블1));
-
-        //when & then
-        assertThatThrownBy(
-                () -> tableGroupService.create(tableGroup)
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("주문 테이블이 빈테이블이 아니면 테이블 그룹을 지정 할 수 없다.")
-    void createFailWithEmptyTableTest() {
-        //given
-        OrderTable orderTable1 = new OrderTable(1L, false);
-        OrderTable orderTable2 = new OrderTable(2L, false);
-        given(orderTableDao.findAllByIdIn(Arrays.asList(orderTable1.getId(), orderTable2.getId()))).willReturn(
-                Arrays.asList(orderTable2, orderTable2));
-
-        //when & then
-        assertThatThrownBy(
-                () -> tableGroupService.create(단체)
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("이미 테이블 그룹에 속해 있으면 테이블 그룹을 지정 할 수 없다.")
-    void createFailWithTableGroupTest() {
-        //given
-        TableGroup tableGroup = new TableGroup(1L);
-        OrderTable orderTable1 = new OrderTable(1L, tableGroup, true);
-        OrderTable orderTable2 = new OrderTable(2L, tableGroup, true);
-        tableGroup.setOrderTables(Arrays.asList(orderTable1, orderTable2));
-
-        given(orderTableDao.findAllByIdIn(Arrays.asList(orderTable1.getId(), orderTable2.getId()))).willReturn(
-                Arrays.asList(orderTable1, orderTable2));
 
         //when & then
         assertThatThrownBy(
